Move JS for tags filters to app/javascript - webpack is needed for latest ES6 features
