+++ app/javascript/packs/application.js
@@ -5,5 +5,8 @@ import 'select2/dist/css/select2.css';
 import { initSelect2 } from '../components/init_select2';
 initSelect2();
 
-import { initMapChoice } from "../components/tag_choice"
-initMapChoice();
+import { initTagsChoice } from "../components/init_tags_choice"
+initTagsChoice();
+
+import { initTagsFilters } from "../components/init_tags_filters"
+initTagsFilters();
